Updated comments.
