+++ src/App.scss
@@ -2,6 +2,6 @@
   background-color: #b91213;
   color: #fff;
 }
-a{
+a {
   text-decoration: none;
 }
