+++ src/components/card-list/CardList.module.scss
@@ -9,6 +9,5 @@
 @media screen and (max-device-width: 1300px) {
 & {
  max-width: 900px;
-//  width: 900px;
 }}
 }
