--- src/components/main-button/main-button.styles.scss
@@ -1,35 +1,0 @@
-.button-container{
-  padding: 18px 5px;
-  width: 230px;
-  text-align: center ;
-  border-radius: 29px;
-  color: #ffffff;
-  border: none;
-  font-size: 18px;
-  text-transform: uppercase;
-  margin: 5px auto;
-  justify-content: center;
-@media screen and (max-device-width: 768px) {
-  & {
-    font-size: 16px;
-  }
-}
-@media screen and (max-device-width: 400px) {
-  & {
-    font-size: 13px;
-  }
-}
-
-}
-
-.log-in{
-  background-color: rgb(31, 62, 81);
-}
-.base{
-  
-  background: linear-gradient(#67747b, #3d4549);
-  
-}
-.len-maxed{
-   background: linear-gradient(#F80705, #7B1E1D);
-}
