--- src/components/return-button/return-button.module.scss
@@ -1,25 +1,0 @@
-
-.return-button {
-  display: flex;
-  margin: 50px 10px;
-  border-radius: 29px;
-  color: #ffffff;
-  border: none;
-  background: linear-gradient(#f80705, #7b1e1d);
-  font-size: clamp(0.9rem, 3.5vw, 1.1rem);
-  text-transform: uppercase;
-  font-family: Montserrat;
-  justify-content: center;
-  align-items: center;
-  width: 350px;
-  width: min(350px,70%);
-  height: 58px;
-  overflow-x: auto;
-
-.padding {
-  padding-right: 5px;
-}
-
-
-
-}
