+++ src/container/leaderboard/leaderboard.module.scss
@@ -1,45 +1,41 @@
-.main-container{
-  background: url('../../assets/background-images/leaderboard.png') no-repeat center fixed;
+.main-container {
+  background: url('../../assets/background-images/leaderboard.png') no-repeat
+    center fixed;
   background-size: cover;
   display: flex;
   flex-direction: column;
   min-height: 100vh;
 
-  
   .glassyBg {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  align-self: center;
-  border-radius: 10px;
-  margin: 50px 20px;
-  width: min(1200px,90%);
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    align-self: center;
+    border-radius: 10px;
+    margin: 50px 20px;
+    width: min(1200px, 90%);
 
-  background: rgba(17, 52, 70, 0.53);
-  backdrop-filter: blur(50px) opacity(1) brightness(-50%);
-  box-shadow: 0 4px 30px rgba(0, 0, 0, 0.2);
+    background: rgba(17, 52, 70, 0.53);
+    backdrop-filter: blur(50px) opacity(1) brightness(-50%);
+    box-shadow: 0 4px 30px rgba(0, 0, 0, 0.2);
 
-  
-.heading {
-  margin-top: 80px;
-  margin-bottom: 20px;
-  color: #fff;
-  font-family: montserrat;
-  
-  font-size: clamp(1.5rem, 3.5vw, 2.1rem);
-  text-align: left;
-  margin-left: 5%;
-  align-self: flex-start;
-  
+    .heading {
+      margin-top: 80px;
+      margin-bottom: 20px;
+      color: #fff;
+      font-family: montserrat;
 
-}
-.main-table {
-  border: 1px solid rgb(255, 255, 255);
-  width: 90%;
-  color: #fff;
-  overflow: auto;
-
-}
-}
+      font-size: clamp(1.5rem, 3.5vw, 2.1rem);
+      text-align: left;
+      margin-left: 5%;
+      align-self: flex-start;
+    }
+    .main-table {
+      border: 1px solid rgb(255, 255, 255);
+      width: 90%;
+      color: #fff;
+      overflow: auto;
+    }
+  }
 }
