+++ src/container/login/login.jsx
@@ -1,15 +1,16 @@
-import Input from 'components/input/input';
-import MainButton from 'components/main-button/main-button';
+import Input from 'components/input/Input';
+import MainButton from 'shared/components/button/Button';
 import logo1 from 'assets/logo/logo1.png';
 import logo2 from 'assets/logo/logo2.png';
 import { ReactComponent as Thumb } from 'assets/thumb.svg';
 import { useEffect, useState } from 'react';
 import { Link, useNavigate } from 'react-router-dom';
-import { useAuthContext } from 'context/AuthContainer';
-import styles from './login.module.scss';
+import { useAuthContext } from 'context/auth-context/AuthContainer';
+import styles from './Login.module.scss';
+import { MAIN_ROUTE } from 'container/routes/constants';
+
 const Login = () => {
   const { isAuthenticated, setLogin } = useAuthContext();
-  console.log('login page', isAuthenticated);
 
   const navigate = useNavigate();
   const [isLenMaxed, setIsLenMaxed] = useState(false);
@@ -24,8 +25,7 @@ const Login = () => {
   const onClickHandler = () => {
     if (isLenMaxed) {
       setLogin();
-      console.log('login button', isAuthenticated);
-      navigate('/mainMenu');
+      navigate(MAIN_ROUTE);
     }
   };
   return (
@@ -39,17 +39,19 @@ const Login = () => {
         <p className={styles['desc']}>
           Enter your game code below to get started
         </p>
+        <div>
+          <Input placeholder={'Game Code'} onChangeHandler={onChangeHandler} />
+          <div className={styles['align-button']}>
+            <MainButton name='test connection' />
 
-        <Input placeholder={'Game Code'} onChangeHandler={onChangeHandler} />
-        <div className={styles['align-button']}>
-          <MainButton name='test connection' />
-          <MainButton
-            buttonType={'log-in'}
-            name={'login'}
-            image={<Thumb />}
-            className={`${isLenMaxed ? 'len-maxed' : ''}`}
-            onClickHandler={onClickHandler}
-          />
+            <MainButton
+              buttonType={'log-in'}
+              name={'login'}
+              rightImage={<Thumb />}
+              className={`${isLenMaxed ? 'len-maxed' : ''}`}
+              onClickHandler={onClickHandler}
+            />
+          </div>
         </div>
       </div>
       <div className={styles['footer']}>
