+++ src/container/main-menu/mainMenu.jsx
@@ -1,14 +1,13 @@
 import login from 'assets/login.png';
 import connected from 'assets/connected.png';
-import { NavButton } from 'components/nav-button/nav-button';
-import { Navigation } from 'container/navigation/navigation';
-import { CardList } from 'components/card-list/card-list';
-import styles from './mainMenu.module.scss';
-import { useAuthContext } from 'context/AuthContainer';
+import { NavButton } from 'components/nav-button/NavButton';
+import { Navigation } from 'container/navigation/Navigation';
+import { CardList } from 'components/card-list/CardList';
+import styles from './MainMenu.module.scss';
+import { useAuthContext } from 'context/auth-context/AuthContainer';
 
 const MainMenu = () => {
   const { isAuthenticated } = useAuthContext();
-  console.log(isAuthenticated);
   return (
     <div className={styles['main-container']}>
       <Navigation />
