+++ src/container/navigation/navigation.jsx
@@ -1,11 +1,11 @@
 import logo1 from 'assets/logo/logo1.png';
 import logo2 from 'assets/logo/logo2.png';
-import login from 'container/login/login';
-import connected from 'assets/connected.png';
-import { NavButton } from 'components/nav-button/nav-button';
+import loginIcon from 'assets/login.png';
+import connectedIcon from 'assets/connected.png';
+import { NavButton } from 'components/nav-button/NavButton';
 import { useNavigate } from 'react-router-dom';
-import { useAuthContext } from 'context/AuthContainer';
-import './navigation.styles.scss';
+import { useAuthContext } from 'context/auth-context/AuthContainer';
+import styles from './Navigation.module.scss';
 
 export const Navigation = () => {
   const navigate = useNavigate();
@@ -16,20 +16,20 @@ export const Navigation = () => {
   };
 
   return (
-    <div className='nav'>
-      <div className='logo'>
+    <div className={styles['nav']}>
+      <div className={styles['logo']}>
         <img src={logo1} alt='experios logo' />
         <img src={logo2} alt='virtual escape logo' />
       </div>
-      <div className='icon'>
-        <i className='fas fa-bars'></i>
+      <div className={styles['icon']}>
+        <i className={styles['fas fa-bars']}></i>
       </div>
 
-      <div className='pages'>
-        <NavButton buttonName={'6 CONNECTED'} imageLink={connected} />
+      <div className={styles['pages']}>
+        <NavButton buttonName={'6 CONNECTED'} imageLink={connectedIcon} />
         <NavButton
           buttonName={'LOG OUT'}
-          imageLink={connected}
+          imageLink={loginIcon}
           onClickHandler={() => onClickHandler('/')}
         />
       </div>
