+++ src/container/routes/PrivateRoutes.jsx
@@ -1,11 +1,10 @@
 import { Outlet, Navigate } from 'react-router-dom';
-import { useAuthContext } from 'context/AuthContainer';
+import { useAuthContext } from 'context/auth-context/AuthContainer';
 import { useContext } from 'react';
 
 const PrivateRoutes = () => {
   const { isAuthenticated } = useAuthContext();
-  console.log('authenticated', isAuthenticated);
-  return isAuthenticated ? <Outlet /> : <Navigate to='/' />;
+  return isAuthenticated ? <Outlet /> : <Navigate to='/login' />;
 };
 
 export default PrivateRoutes;
