+++ src/container/routes/constants.js
@@ -1,0 +1,3 @@
+export const MAIN_ROUTE = '/';
+export const LEADERBOARD_ROUTE = '/leaderboard';
+export const LOGIN_ROUTE = '/login';
