+++ src/data/table/header.js
@@ -1,0 +1,7 @@
+export const headerNames = [
+  'Rank',
+  'Team Name',
+  'Hints',
+  'Incorrect',
+  'Results',
+];
