+++ src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { BrowserRouter } from 'react-router-dom';
-import AuthContextContainer from './context/AuthContainer';
+import AuthContextContainer from './context/auth-context/AuthContainer';
 import './index.scss';
 import App from './App';
