+++ src/index.scss
@@ -1,6 +1,5 @@
 body {
-  scrollbar-gutter: stable ;
-
+  scrollbar-gutter: stable;
   margin: 0px 0px;
   font-family: -apple-system, BlinkMacSystemFont, montserrat, 'Segoe UI',
     'Roboto', 'Oxygen', 'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans',
