+++ src/shared/components/button/Button.module.scss
@@ -1,0 +1,51 @@
+.button-container {
+  padding: 18px 5px;
+  width: 230px;
+  text-align: center;
+  border-radius: 29px;
+  color: #ffffff;
+  border: none;
+  font-size: 18px;
+  text-transform: uppercase;
+  margin: 5px auto;
+  justify-content: center;
+
+  @media screen and (max-device-width: 768px) {
+    & {
+      font-size: 16px;
+    }
+  }
+  @media screen and (max-device-width: 400px) {
+    & {
+      font-size: 13px;
+    }
+  }
+}
+
+.log-in {
+  background-color: rgb(31, 62, 81);
+}
+.base {
+  background: linear-gradient(#67747b, #3d4549);
+}
+.len-maxed {
+  background: linear-gradient(#f80705, #7b1e1d);
+}
+
+.return-to-main {
+  display: flex;
+  margin: 50px 10px;
+  border-radius: 29px;
+  color: #ffffff;
+  border: none;
+  background: linear-gradient(#f80705, #7b1e1d);
+  font-size: clamp(0.9rem, 3.5vw, 1.1rem);
+  text-transform: uppercase;
+  font-family: Montserrat;
+  justify-content: center;
+  align-items: center;
+  width: 350px;
+  width: min(350px, 70%);
+  height: 58px;
+  overflow-x: auto;
+}
