+++ src/shared/components/card/Card.jsx
@@ -1,14 +1,15 @@
-import './card.styles.scss';
 import { useNavigate } from 'react-router-dom';
+import styles from './Card.module.scss';
+
 export const Card = ({ name, image, onClickNavigate }) => {
   const navigate = useNavigate();
   const onClickHandler = () =>
     onClickNavigate ? navigate(onClickNavigate) : null;
 
   return (
-    <div className={`card-container `} onClick={onClickHandler}>
+    <div className={styles['card-container']} onClick={onClickHandler}>
       <img src={image} />
-      <h2 className={'text-container '}>{name}</h2>
+      <h2 className={styles['text-container']}>{name}</h2>
     </div>
   );
 };
