+++ src/shared/components/icon/Icon.module.scss
@@ -1,0 +1,14 @@
+.icon-container {
+  border-radius: 50%;
+  width: 25px;
+  height: 25px;
+  font-weight: bold;
+  position: absolute;
+  right: 20px;
+  bottom: 13px;
+  background-color: #b91213;
+  color: #ffffff;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
