+++ src/shared/components/table-header/TableHeader.module.scss
@@ -1,4 +1,3 @@
-
 th {
   padding: 15px 5px;
   font-weight: normal;
