--- src/shared/components/table-header/table-header.jsx
@@ -1,13 +1,0 @@
-import styles from './table-header.module.scss';
-
-export const TableHeader = () => {
-  return (
-    <tr>
-      <th>Rank</th>
-      <th>Team Name</th>
-      <th>Hints</th>
-      <th>Incorrect</th>
-      <th>Results</th>
-    </tr>
-  );
-};
