+++ src/shared/components/table/table.module.scss
@@ -1,6 +1,5 @@
-
 table {
   border-collapse: collapse;
   font-family: montserrat;
-  width: 100%
+  width: 100%;
 }
